Fix for : below issue: twitter4j-core-4.0.2.jar, InvalidPackage, Priority: High Invalid package reference in library; not included in Android: java.lang.management. Referenced from twitter4j.TwitterAPIMonitor.
This check scans through libraries looking for calls to APIs that are not included in Android. When you create Android projects, the classpath is set up such that you can only access classes in the API packages that are included in Android. However, if you add other projects to your libs/ folder, there is no guarantee that those .jar files were built with an Android specific classpath, and in particular, they could be accessing unsupported APIs such as java.applet. This check scans through library jars and looks for references to API packages that are not included in Android and flags these. This is only an error if your code calls one of the library classes which wind up referencing the unsupported package.

Reviewed by: naveen
comments: NA