Change Description -  Made the fix in my earlier crash.
1) If UR is not logged in & click on IAP then it crashes. 100% occurrence.

Now reverting those changes.  As per confirmation from Rama (PO, Amalgam) there is a flow change. When user is not logged in then he is not allowed to enter into RefApp_Home+Screen. So this bug is on hold.

User Story/Defect ID  - 26000

Reviewer :  Team will review it.

Comments :   NA