Merged PR 5790: Fixes data sync to recover from error
DSC component would, when pulling data from OBE, save the error to report to the API user.
In case of the auth token expiring, DSC recovers from this by requesting a new token.
The original error (token expired) was held on to, so even after receiving a new token, the error would remain.
These changes fix that behaviour, resetting the error after it was posted in a callback.

Related work items: #122313