Merged PR 5839: Null safety in DSC for updating moments
The issue reports a NullPointerException in some cases.
A null safety was built in to prevent a crash in those cases.

We were unable to reproduce, but by manipulating some tests, it crashed indeed.
Root cause for the null value was not found, but prevented crashing with this check.

Related work items: #121399