Merged PR 8048: Merge feature/plf/gitignore_cleanup to develop
Performed a clean up of the huge forest of .gitignore files.
Moved majority of rules to global .gitignore file and removed them from the local file.
If the resulting local file is empty, then it is deleted.

Please check for mistakes in general and in the context of your components in specific.